+++ packages/design/narduk-charts/site/views/examples/BarExample.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import { NardukBarChart } from 'narduk-charts'
-import type { ChartSeries } from 'narduk-charts'
+import { NardukBarChart } from '@narduk-enterprises/narduk-charts'
+import type { ChartSeries } from '@narduk-enterprises/narduk-charts'
 import ExamplePage from '../../components/ExamplePage.vue'
 
 const labels = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun']