+++ packages/design/narduk-charts/site/views/examples/StreamingExample.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { onMounted, onUnmounted, ref, computed } from 'vue'
-import { NardukLineChart, useStreamingSeries } from 'narduk-charts'
+import { NardukLineChart, useStreamingSeries } from '@narduk-enterprises/narduk-charts'
 import ExamplePage from '../../components/ExamplePage.vue'
 
 const WINDOW = 48